[FEAT]: DIContainer를 Core로 이동 및 SceneDelegate에서 의존성 주입
