+++ App/App/SceneDelegate.swift
@@ -5,15 +5,20 @@
 //  Created by jung on 11/4/24.
 //
 
-import UIKit
+import Core
+import Data
 import Feature
+import Interfaces
+import P2PSocket
+import UIKit
+import UseCase
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 	var window: UIWindow?
 	
 	func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
 		guard let windowScene = (scene as? UIWindowScene) else { return }
-        DIContainer.shared.registerDependency()
+        registerDependency()
 
 		let window = UIWindow(windowScene: windowScene)
         let connectionViewController = ConnectionViewController(
@@ -23,4 +28,78 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 		self.window = window
 		window.makeKeyAndVisible()
 	}
+}
+
+extension SceneDelegate {
+    func registerDependency() {
+        registerP2PSocket()
+        registerRepository()
+        registerUseCase()
+        registerViewModel()
+    }
+
+    func registerP2PSocket() {
+        DIContainer.shared.register(
+            type: SocketProvidable.self,
+            instance: SocketProvider()
+        )
+    }
+
+    func registerRepository() {
+        DIContainer.shared.register(
+            type: BrowsingUserRepositoryInterface.self,
+            instance: BrowsingUserRepository(
+                socketProvider: DIContainer.shared.resolve(type: SocketProvidable.self)
+            )
+        )
+
+        DIContainer.shared.register(
+            type: ConnectedUserRepositoryInterface.self,
+            instance: ConnectedUserRepository(
+                socketProvider: DIContainer.shared.resolve(type: SocketProvidable.self)
+            )
+        )
+
+        DIContainer.shared.register(
+            type: SharingVideoRepositoryInterface.self,
+            instance: SharingVideoRepository(
+                socketProvider: DIContainer.shared.resolve(type: SocketProvidable.self)
+            )
+        )
+    }
+
+    func registerUseCase() {
+        DIContainer.shared.register(
+            type: BrowsingUserUseCaseInterface.self,
+            instance: BrowsingUserUseCase(
+                repository: DIContainer.shared.resolve(type: BrowsingUserRepositoryInterface.self)
+            )
+        )
+
+        DIContainer.shared.register(
+            type: ConnectedUserUseCaseInterface.self,
+            instance: ConnectedUserUseCase(
+                repository: DIContainer.shared.resolve(type: ConnectedUserRepositoryInterface.self)
+            )
+        )
+    }
+
+    func registerViewModel() {
+        DIContainer.shared.register(
+            type: ConnectionViewModel.self,
+            instance: ConnectionViewModel(
+                usecase: DIContainer.shared.resolve(type: BrowsingUserUseCaseInterface.self)
+            )
+        )
+
+        DIContainer.shared.register(
+            type: GroupInfoViewModel.self,
+            instance: GroupInfoViewModel()
+        )
+
+        DIContainer.shared.register(
+            type: MultipeerVideoListViewModel.self,
+            instance: MultipeerVideoListViewModel()
+        )
+    }
 }
